FUSE-901: added ThreadEvents class
this is a thin wrapper around Guava EventBus that will be used to send events
like PR creation internally within the DVCS connector.

the synchronisation code will then collect these events and propagate them to
the JIRA EventPublisher at the end of the sync process. if can also choose to
ignore events raised during "hard" sync as these should not be propagated.